+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Fixed English translations of order statuses created by the module
 - Fixed HRX block in Order view on Prestashop 1.7.0-1.7.6
 - Fixed page title in Prestashop Order page
+- Fixed Order status change to "New HRX shipment"
+
+### Improved
+- Reduced the number of requests to the database when changing the order status
 
 ## [1.2.3] - Fix track number
 ### Fixed